air/requirements.txt
ultralytics>=8.2.0
opencv-python>=4.9.0.80
numpy>=1.26.0
deep-sort-realtime>=1.3.2
filterpy>=1.4.5
pyserial>=3.5
